+++ src/app/app.module.ts
@@ -2,8 +2,17 @@ import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule } from '@angular/core';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { FormsModule } from '@angular/forms';
-import { MatIconModule, MatButtonModule, MatMenuModule, MatToolbarModule, MatSelectModule, MatSidenavModule } from '@angular/material';
+import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import {
+  MatButtonModule,
+  MatFormFieldModule,
+  MatIconModule,
+  MatInputModule,
+  MatMenuModule,
+  MatToolbarModule,
+  MatSelectModule,
+  MatSidenavModule
+} from '@angular/material';
 import { PipThemesModule } from 'pip-webui2-themes';
 
 import { AppComponent } from './app.component';
@@ -18,12 +27,15 @@ import { TestModule } from './test/test.module';
     BrowserAnimationsModule,
     FlexLayoutModule,
     FormsModule,
-    MatIconModule,
     MatButtonModule,
+    MatFormFieldModule,
+    MatIconModule,
+    MatInputModule,
     MatMenuModule,
     MatToolbarModule,
     MatSelectModule,
     MatSidenavModule,
+    ReactiveFormsModule,
 
     TestModule,
     PipThemesModule.forRoot()